Change image for "check mark" bullet points
The feedback I got was that the previous style--a filled box with a
white check mark--looked too much like an interactable check box. The
tendency was to click on them to select exactly the topics people were
interested in.

I switched it over to just a check mark, without the accompanying box.