Refactor subscribe form into an include
This subscribe form will be used in many different pages, including the
archive pages. While it would be nice to be able to customize this form
based on where it is included, it should be sufficient to include it
without modification, at least for now.

Also refactor the corresponding CSS into a partial that can be
`import`ed by SASS into multiple stylesheets.