Fix image path on 404 page
Should have used an absolute path, as it turns out the Jekyll server (or
really whatever server is ultimately serving the website) presents the
HTML for the 404 page under the same URL that's missing. That means, if
the missing page has a path like "nested/directory/missing.html", the
image has te be resolved relative to `nested/directory`.