Edition: solving iterator interview problems with recursion
This is the first edition that has embedded code. Normally, I would
write about a topic like this in my personal website and link to it from
the newsletter. This time, however, the article is self-contained,
related directly to hiring, and I don't have much commentary to add.

To support syntax highlighting, I added the Github syntax highlighting
CSS. In the future, I may include this CSS file conditionally based on
whether the edition needs it. For now, I just added it.