Removed references to LinkedIn Group
The group was always an experiment, and it didn't really take off. So
I'm retiring it and removing all references to the group.

Previously, the bottom of each post pointed readers to my Twitter. I
kept the Twitter link on the index page, but I didn't go back to
prominently linking the Twitter page because Twitter has kind of gone
downhill. In the future, I can imagine bringing such a link back if I
get on Mastodon.